move input section into the components folder
