+++ src/packages/core/section/components/index.ts
@@ -1,1 +1,2 @@
+export * from './input-section/index.js';
 export * from './ref-section/ref-section.element.js';
