+++ src/packages/core/section/components/input-section/input-section.context.ts
@@ -1,6 +1,6 @@
-import type { UmbSectionItemModel } from '../repository/index.js';
-import { UMB_SECTION_ITEM_REPOSITORY_ALIAS } from '../repository/index.js';
-import { UMB_SECTION_PICKER_MODAL } from '../section-picker-modal/section-picker-modal.token.js';
+import type { UmbSectionItemModel } from '../../repository/index.js';
+import { UMB_SECTION_ITEM_REPOSITORY_ALIAS } from '../../repository/index.js';
+import { UMB_SECTION_PICKER_MODAL } from '../../section-picker-modal/section-picker-modal.token.js';
 import { UmbPickerInputContext } from '@umbraco-cms/backoffice/picker-input';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
