+++ src/packages/core/section/index.ts
@@ -1,10 +1,9 @@
-export * from './section-main/index.js';
-export * from './section-sidebar/index.js';
-export * from './section-sidebar-context-menu/index.js';
-export * from './section-sidebar-menu/index.js';
-export * from './section-sidebar-menu-with-entity-actions/index.js';
-export * from './section-default.element.js';
-export * from './section.context.js';
-export * from './input-section/index.js';
-export * from './section-picker-modal/section-picker-modal.token.js';
 export * from './components/index.js';
+export * from './section-default.element.js';
+export * from './section-main/index.js';
+export * from './section-picker-modal/section-picker-modal.token.js';
+export * from './section-sidebar-context-menu/index.js';
+export * from './section-sidebar-menu-with-entity-actions/index.js';
+export * from './section-sidebar-menu/index.js';
+export * from './section-sidebar/index.js';
+export * from './section.context.js';
